🔨🤖 Rename to food_supply_chain_fbs and follow the prototype's assumptions: fat ceiling, net processing, rounding into residuals
